Major update: 1. added several special folders 2. added the deletion of a deck once no files is in it (make sure that you do not add any note to it directly via anki)

+++ code/anki_importer.py
@@ -15,7 +15,7 @@ def importer(my_files_catalog):
 	delete_empty_decks()
 	
 def importer_to_anki(file):
-	if file.get_file_root_folder() == "Trash":
+	if file.get_file_root_folder() == settings.get_settings_by_name("trash folder"):
 		uid = file.get_file_uid()
 		note_list = mw.col.find_notes(uid)
 		if len(note_list) > 0:
@@ -26,18 +26,26 @@ def importer_to_anki(file):
 						mw.col.remNotes([single_note_id])
 				except KeyError:
 					pass
-#		os.remove(file.get_file_full_path())
+	elif file.get_file_root_folder() == settings.get_settings_by_name("archive folder") or file.get_file_root_folder() == settings.get_settings_by_name("templates folder"):
+		uid = file.get_file_uid()
+		note_list = mw.col.find_notes(uid)
+		if len(note_list) > 0:
+			for single_note_id in note_list:
+				single_note = mw.col.getNote(single_note_id)
+				if single_note["UID"] == uid:
+					mw.col.remNotes([single_note_id])
 	else:
 		deck_id = mw.col.decks.id(file.get_deck_name())
 		mw.col.decks.select(deck_id)
 		card_model = mw.col.models.byName("Obsidianki4")
-		note_list = mw.col.find_notes(file.get_file_uid())
+		uid = file.get_file_uid()
+		note_list = mw.col.find_notes(uid)
 		found_exisiting_file = False
 		if len(note_list) > 0:
 			for single_note_id in note_list:
 				single_note = mw.col.getNote(single_note_id)
 				if single_note.model() == card_model:
-					if single_note["UID"] == file.get_file_uid():
+					if single_note["UID"] == uid:
 						if file.get_cloze_or_basic():
 							single_note["Cloze"] = file.get_file_content()
 							single_note["Text"] = ""
@@ -51,9 +59,12 @@ def importer_to_anki(file):
 						single_note.tags = []
 						for tag in file.get_tags():
 							single_note.tags.append(tag)
-						
-						card_ids = mw.col.card_ids_of_note(single_note_id)
-						mw.col.set_deck(card_ids, deck_id)
+						try:
+							card_ids = mw.col.card_ids_of_note(single_note_id)
+							mw.col.set_deck(card_ids, deck_id)
+						except AttributeError:
+							card_ids = mw.col.find_cards(uid)
+							mw.col.decks.setDeck(card_ids, deck_id)
 						single_note.flush()
 						found_exisiting_file = True
 		if not found_exisiting_file:
@@ -68,7 +79,7 @@ def importer_to_anki(file):
 				else:
 					note_object["Cloze"] = "{{c1::}}"
 					note_object["Text"] = file.get_file_content()
-				note_object["UID"] = file.get_file_uid()
+				note_object["UID"] = uid
 				back_extra = "Source: " + file.get_file_name_with_url()
 				note_object["Back Extra"] = back_extra
 				for tag in file.get_tags():
@@ -83,15 +94,28 @@ def delete_empty_decks():
 	for name_and_id in names_and_ids:
 		# I could not find what type this object is, so the only way for me to do it now is to use the string.
 		name_and_id_segments = str(name_and_id).split("\n")
-		id = int(name_and_id_segments[0].split(": ")[1])
-		if id != 1 and mw.col.decks.card_count(id, True) == 0:
-			mw.col.decks.rem(id, True, True)
+		deck_id= int(name_and_id_segments[0].split(": ")[1])
+		
+		if deck_has_cards(deck_id):
+			mw.col.decks.rem(deck_id, True, True)
 			
 def empty_trash():
 	path = settings.get_settings_by_name("vault path")
 	# TODO: Add this to settings
-	trash_can_path = path + "/" + "Trash"
-	trash_directories = os.listdir(trash_can_path)
-	for trash_directory in trash_directories:
-		trash_directory_path = trash_can_path + "/" + trash_directory
-		shutil.rmtree(trash_directory_path)
+	if settings.get_settings_by_name("trash folder") != "":
+		trash_can_path = path + "/" + settings.get_settings_by_name("trash folder")
+		trash_directories = os.listdir(trash_can_path)
+		for trash_directory in trash_directories:
+			trash_directory_path = trash_can_path + "/" + trash_directory
+			shutil.rmtree(trash_directory_path)
+		
+def deck_has_cards(deck_id):
+	if deck_id != 1:
+		try:
+			if mw.col.decks.card_count(deck_id, True) == 0:
+				return True
+		except AttributeError:
+			cids = mw.col.decks.cids(deck_id, True)
+			if len(cids) == 0:
+				return True
+	return False
